actionRepaier.tsを作成する
